backend/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended>=4.0.0
Flask-CORS>=4.0.0
Werkzeug>=2.0.0
Pillow>=9.0.0
psutil>=5.8.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
gunicorn>=20.0.0